# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation API Documentation",
        "description": "API documentation for the Translation Management System",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "API Server"
        },
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get list of translations",
                "description": "Returns list of all translations with optional filtering",
                "operationId": "getTranslationsList",
                "parameters": [
                    {
                        "name": "language_id",
                        "in": "query",
                        "description": "Filter by language ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "group",
                        "in": "query",
                        "description": "Filter by translation group",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filter by translation tag",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Filter by translation key",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "language_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "key": {
                                                "type": "string",
                                                "example": "welcome.message"
                                            },
                                            "value": {
                                                "type": "string",
                                                "example": "Welcome to our application"
                                            },
                                            "group": {
                                                "type": "string",
                                                "example": "messages"
                                            },
                                            "tags": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "frontend"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Frontend translations"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a new translation",
                "description": "Creates a new translation and returns the created resource",
                "operationId": "storeTranslation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "language_id",
                                    "key",
                                    "value",
                                    "group"
                                ],
                                "properties": {
                                    "language_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "welcome.message"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "Welcome to our application"
                                    },
                                    "group": {
                                        "type": "string",
                                        "example": "messages"
                                    },
                                    "tags": {
                                        "description": "Array of tag IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "language_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "welcome.message"
                                        },
                                        "value": {
                                            "type": "string",
                                            "example": "Welcome to our application"
                                        },
                                        "group": {
                                            "type": "string",
                                            "example": "messages"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "frontend"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/translations/by-key": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get translations by key",
                "description": "Returns translations for a specific key across all languages",
                "operationId": "getTranslationsByKey",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Translation key to search for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "query",
                        "description": "Translation group",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "language_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "key": {
                                                "type": "string",
                                                "example": "welcome.message"
                                            },
                                            "value": {
                                                "type": "string",
                                                "example": "Welcome to our application"
                                            },
                                            "group": {
                                                "type": "string",
                                                "example": "messages"
                                            },
                                            "language": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "en"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "English"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}